Moved the URL list to individual apps via callbacks

+++ handlers/admin.php
@@ -7,8 +7,10 @@ $page->layout = 'admin';
 $page->title = 'Google Sitemaps';
 
 if (isset ($_GET['reset'])) {
-	unlink ('cache/sitemap.xml');
-	$this->add_notification (i18n_get ('sitemap.xml reset.'));
+	if (file_exists ('cache/sitemap.xml')) {
+		unlink ('cache/sitemap.xml');
+	}
+	$this->add_notification (__ ('sitemap.xml reset.'));
 	$this->redirect ('/sitemap.xml/admin');
 }
